+++ CHANGELOG.md
@@ -27,3 +27,4 @@
 - Mise en place du loader
 - Mise en place des notifications
 - Mise en place du menu
+- Mise en place du header